Fixing errors in modules
