+++ DungeonProject/DungeonProject/MathModule.py
@@ -32,6 +32,7 @@ class MathModule:
             }.get(angle, math.cos(math.radians(angle)))
 
     # Returns the eight corner points of the given BB
+    @staticmethod
     def getBBpoints(bb):
         vectors = [
             [[bb[1][0][0] * cos(bb[0][3]), 0, bb[1][0][0] * sin(bb[0][3])], 
@@ -54,6 +55,7 @@ class MathModule:
 
     # Adds an infinite amount of vectors
     # PRE: all vectors must have the same dimension (behaviour undefined if not true)
+    @staticmethod
     def addVectors(vectors):
         result = vectors[0]
         for v in vectors[1:]:
@@ -62,6 +64,7 @@ class MathModule:
         return result
 
     # Creates the AABB that contains all points
+    @staticmethod
     def AABB(points):
         points = zip(*points)
         min_x = min(points[0])
@@ -69,6 +72,21 @@ class MathModule:
         min_y = min(points[1])
         max_y = max(points[1])
         return (min_x, min_y, max_x - min_x, max_y - min_y)
+
+    # Projects a point onto a vertical origin-passing plane given by the angle
+    # Returns the projected point in the local coordinates of the plane
+    @staticmethod
+    def projectPointOntoPlane(angle, point):
+        # Computes the projection of the point onto the normal vector
+        normal = [-sin(angle), 0, cos(angle)]
+        dotProduct = point[0] * normal[0] + point[2] * normal[2]
+        projection = [v * dotProduct for v in normal]
+
+        # Substract the projection the the original point to obtain the projected point
+        projected = [point[i] - projection[i] for i in range(len(point))]
+
+        # Change the point to return it in the local coordinates of the plane
+        return [projected[0] * cos(-angle) - projected[2] * sin(-angle), projected[1]]
